requirements.txt
alpaca==1.0.0
langchain==0.3.26
langchain_chroma==0.2.4
langchain_core==0.3.66
langchain_google_genai==2.1.5
pandas==2.3.0
python-dotenv==1.1.1
python_binance==1.0.29
Requests==2.32.4
